E2E pipeline updates
Previously there was a wercker pipeline per e2e test, thus the test
parallelism was achieved thorugh wercker. However, this meant that
adding an e2e test was painful, as we had to add the test, update the
wercker yaml and configure the new piepline to run in the wercker UI.

This patch aims to implement the test parallelism inside gotest instead.
We can then have just 2 piplines, one for running the upgrade test (as
it needs it's own operator) and the other for running all of the other
e2e test, thus adding a new test should just be a case of adding the
test code.

Note: The upgrade pipeline runs with non-buffered logs. This is ok
as there is only a single test in this suite. Also, this prevents
the test timing out in wercker due to lack of activity on stdout.