Fix tutorial bugs
This fixes an inconsistency where you add the mysql-agent sa to
"my-namespace" but then proceed to add a database to "default".

Set apiVersion to the correct value of "mysql.oracle.com/v1alpha1"

---

I make this contribution available under the Apache License 2.0. I do
not, and will not, sign the Oracle CLA as I do not feel it is worth the
time nor money to have my attorney review it. Based on Oracle's past
behavior with regard to the Linux community, I do not trust an Oracle
agreement to have the best interest of the community as its intent
thus will not sign without legal review.